+++ lib/ac-api/src/main/java/com/webcodepro/applecommander/storage/DiskFactory.java
@@ -19,9 +19,11 @@
  */
 package com.webcodepro.applecommander.storage;
 
-import org.applecommander.codec.Nibble53Disk525Codec;
-import org.applecommander.codec.Nibble62Disk525Codec;
+import org.applecommander.device.nibble.Nibble53Disk525Codec;
+import org.applecommander.device.nibble.Nibble62Disk525Codec;
 import org.applecommander.device.*;
+import org.applecommander.device.nibble.DiskMarker;
+import org.applecommander.device.nibble.NibbleTrackReaderWriter;
 import org.applecommander.hint.Hint;
 import org.applecommander.image.NibbleImage;
 import org.applecommander.image.WozImage;
