+++ lib/ac-api/src/main/java/org/applecommander/device/nibble/NibbleTrackReaderWriter.java
@@ -17,14 +17,23 @@
  * with this program; if not, write to the Free Software Foundation, Inc.,
  * 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  */
-package org.applecommander.codec;
+package org.applecommander.device.nibble;
 
 import org.applecommander.capability.CapabilityProvider;
+import org.applecommander.util.Container;
 import org.applecommander.util.DataBuffer;
 
-public interface NibbleDiskCodec extends CapabilityProvider {
-    int encodedSize();
-    int decodedSize();
-    DataBuffer decode(DataBuffer rawData);
-    DataBuffer encode(DataBuffer data);
+public interface NibbleTrackReaderWriter extends CapabilityProvider, Container {
+    /**
+     * The number of physical tracks on this device.
+     */
+    int getTracksOnDevice();
+    /**
+     * Read nibbilized track data.
+     */
+    DataBuffer readTrackData(int track);
+    /**
+     * Write nibbilized track data.
+     */
+    void writeTrackData(int track, DataBuffer data);
 }
