+++ lib/ac-swt-common/src/main/java/com/webcodepro/applecommander/ui/swt/DiskExplorerTab.java
@@ -44,8 +44,9 @@ import com.webcodepro.applecommander.util.Host;
 import com.webcodepro.applecommander.util.StreamUtil;
 import com.webcodepro.applecommander.util.TextBundle;
 import io.github.applecommander.applesingle.AppleSingle;
-import org.applecommander.codec.Nibble62Disk525Codec;
+import org.applecommander.device.nibble.Nibble62Disk525Codec;
 import org.applecommander.device.*;
+import org.applecommander.device.nibble.DiskMarker;
 import org.applecommander.hint.Hint;
 import org.applecommander.image.NibbleImage;
 import org.applecommander.source.DataBufferSource;
